release 2025-04-25

+++ Rakefile
@@ -340,6 +340,8 @@ module TogoID
         return "prepare:cog"
       when /#{OUTPUT_TSV_DIR}interpro/
         return "prepare:interpro"
+      when /#{OUTPUT_TSV_DIR}lipidmaps/
+        return "prepare:lipidmaps"
       when /#{OUTPUT_TSV_DIR}mgi_gene/
         return "prepare:mgi_gene"
       when /#{OUTPUT_TSV_DIR}mgi_genotype/
@@ -509,7 +511,7 @@ end
 
 namespace :prepare do
   desc "Prepare all"
-  task :all => [ :bioproject, :biosample, :cellosaurus, :clinvar, :ensembl, :flybase, :glytoucan, :hmdb, :hgnc, :homologene, :hp_phenotype, :cog, :interpro, :mgi_gene, :mgi_genotype, :ncbigene, :oma_protein, :pmc, :prosite, :reactome, :refseq_protein, :refseq_rna, :rhea, :rnacentral, :sra, :swisslipids, :uniprot, :taxonomy, :zfin ]
+  task :all => [ :bioproject, :biosample, :cellosaurus, :clinvar, :ensembl, :flybase, :glytoucan, :hmdb, :hgnc, :homologene, :hp_phenotype, :cog, :interpro, :lipidmaps, :mgi_gene, :mgi_genotype, :ncbigene, :oma_protein, :pmc, :prosite, :reactome, :refseq_protein, :refseq_rna, :rhea, :rnacentral, :sra, :swisslipids, :uniprot, :taxonomy, :zfin ]
 
   directory INPUT_DRUGBANK_DIR    = "input/drugbank"
   directory INPUT_BIOPROJECT_DIR  = "input/bioproject"
@@ -525,6 +527,7 @@ namespace :prepare do
   directory INPUT_HMDB_DIR        = "input/hmdb"
   directory INPUT_HGNC_DIR        = "input/hgnc"
   directory INPUT_INTERPRO_DIR    = "input/interpro"
+  directory INPUT_LIPIDMAPS_DIR   = "input/lipidmaps"
   directory INPUT_MGI_GENE_DIR    = "input/mgi_gene"
   directory INPUT_MGI_GENOTYPE_DIR    = "input/mgi_genotype"
   directory INPUT_NCBIGENE_DIR    = "input/ncbigene"
@@ -645,10 +648,17 @@ namespace :prepare do
     $stderr.puts "## Prepare input files for FlyBase"
     download_lock(INPUT_FLYBASE_DIR) do
       updated = false
-      sh "wget ftp://ftp.flybase.net/releases/current/precomputed_files/genes/ --no-remove-listing --directory-prefix=#{INPUT_FLYBASE_DIR}"
-      current_filename = `grep -oE "fbgn_fbtr_fbpp_expanded_fb_.*\.tsv\.gz" #{INPUT_FLYBASE_DIR}/.listing`.strip
-      input_file = "#{INPUT_FLYBASE_DIR}/#{current_filename}"
-      if !File.exist?(input_file)
+
+      begin
+        sh "wget ftp://ftp.flybase.net/releases/current/precomputed_files/genes/ --no-remove-listing --directory-prefix=#{INPUT_FLYBASE_DIR}"
+        current_filename = `grep -oE "fbgn_fbtr_fbpp_expanded_fb_.*\.tsv\.gz" #{INPUT_FLYBASE_DIR}/.listing`.strip
+        input_file = "#{INPUT_FLYBASE_DIR}/#{current_filename}"
+      rescue StandardError => e
+        $stderr.puts "Error: wget: #{e.message}"
+        input_file = ""
+      end
+
+      if input_file != "" && !File.exist?(input_file)
         sh "rm -f #{INPUT_FLYBASE_DIR}/fbgn_fbtr_fbpp_expanded_fb_*.tsv.gz"
         input_url = "ftp://ftp.flybase.net/releases/current/precomputed_files/genes/#{current_filename}"
         download_file(INPUT_FLYBASE_DIR, input_url)
@@ -798,6 +808,29 @@ namespace :prepare do
     end
   end
 
+  desc "Prepare required files for LIPID MAPS"
+  task :lipidmaps => INPUT_LIPIDMAPS_DIR do
+    $stderr.puts "## Prepare input files for LIPID MAPS"
+    download_lock(INPUT_LIPIDMAPS_DIR) do
+      updated = false
+      input_file = "#{INPUT_LIPIDMAPS_DIR}/sdf.zip"
+      input_url  = "\"https://www.lipidmaps.org/files/?file=LMSD&ext=sdf.zip\""
+      ## Check only the file size because the timestamp is not available.
+      if check_remote_file_size(input_file, input_url)
+        opts = "--quiet --no-check-certificate"
+        begin
+          sh "wget #{opts} -O #{INPUT_LIPIDMAPS_DIR}/sdf.zip #{input_url}"
+        rescue StandardError => e
+          $stderr.puts "Error: download_file(#{INPUT_LIPIDMAPS_DIR}, #{input_url}): #{e.message}"
+        end
+        sh "unzip -o #{input_file} -d #{INPUT_LIPIDMAPS_DIR}/"
+        sh "awk -f bin/parse_lipidmaps_sdf.awk #{INPUT_LIPIDMAPS_DIR}/structures.sdf > #{INPUT_LIPIDMAPS_DIR}/lipidmaps.tsv"
+        updated = true
+      end
+      updated
+    end
+  end
+
   desc "Prepare required files for MGI gene"
   task :mgi_gene => INPUT_MGI_GENE_DIR do
     $stderr.puts "## Prepare input files for MGI_GENE"
@@ -905,9 +938,15 @@ namespace :prepare do
           updated = true
         end
       end
-      if updated
-        sh "awk -F \"\t\" '$4==\"Ensembl\" && $5~/Ensembl (Vertebrates )?[0-9]+;/ && $1!=\"YEAST\"{print $2}' #{INPUT_OMA_PROTEIN_DIR}/oma-species.txt > #{INPUT_OMA_PROTEIN_DIR}/taxonomy.txt"
+
+      ensembl_taxonomy_file = "#{INPUT_OMA_PROTEIN_DIR}/ensembl_vertebrate_taxonomy.txt"
+      oma_ensembl_taxonomy_file = "#{INPUT_OMA_PROTEIN_DIR}/oma_ensembl_vertebrate_taxonomy.txt"
+      if updated || !File.exist?(ensembl_taxonomy_file)
+        sh "sparql_csv2tsv.sh #{INPUT_ENSEMBL_DIR}/taxonomy.rq https://rdfportal.org/ebi/sparql > #{ensembl_taxonomy_file}"
+        sh "awk -F \"\t\" 'FNR==NR{a[$1]=1;next} a[$3]{print $3}' #{ensembl_taxonomy_file} #{INPUT_OMA_PROTEIN_DIR}/oma-species.txt > #{oma_ensembl_taxonomy_file}"
+        updated = true
       end
+
       updated
     end
   end
